+++ src/command.rs
@@ -3,17 +3,25 @@ use crate::prelude::*;
 #[derive(Debug)]
 pub enum Command {
     ReadInputs(config::Inverter, i16),
-    ReadInput(config::Inverter, i16, i16),
-    ReadHold(config::Inverter, i16, i16),
+    ReadInput(config::Inverter, i16, u16),
+    ReadHold(config::Inverter, i16, u16),
     ReadParam(config::Inverter, i16),
-    SetHold(config::Inverter, i16, i16),
-    ChargeRate(config::Inverter, i16),
-    DischargeRate(config::Inverter, i16),
+    ReadAcChargeTime(config::Inverter, i16),
+    ReadChargePriorityTime(config::Inverter, i16),
+    ReadForcedDischargeTime(config::Inverter, i16),
+    SetHold(config::Inverter, i16, u16),
+    WriteParam(config::Inverter, i16, u16),
+    SetAcChargeTime(config::Inverter, i16, [u8; 4]),
+    SetChargePriorityTime(config::Inverter, i16, [u8; 4]),
+    SetForcedDischargeTime(config::Inverter, i16, [u8; 4]),
+    ChargeRate(config::Inverter, u16),
+    DischargeRate(config::Inverter, u16),
     AcCharge(config::Inverter, bool),
+    ChargePriority(config::Inverter, bool),
     ForcedDischarge(config::Inverter, bool),
-    AcChargeRate(config::Inverter, i16),
-    AcChargeSocLimit(config::Inverter, i16),
-    DischargeCutoffSocLimit(config::Inverter, i16),
+    AcChargeRate(config::Inverter, u16),
+    AcChargeSocLimit(config::Inverter, u16),
+    DischargeCutoffSocLimit(config::Inverter, u16),
 }
 
 impl Command {
@@ -31,10 +39,32 @@ impl Command {
             ReadParam(inverter, register) => {
                 format!("{}/read/param/{}", inverter.datalog(), register)
             }
+            ReadAcChargeTime(inverter, num) => {
+                format!("{}/read/ac_charge/{}", inverter.datalog(), num)
+            }
+            ReadChargePriorityTime(inverter, num) => {
+                format!("{}/read/charge_priority/{}", inverter.datalog(), num)
+            }
+            ReadForcedDischargeTime(inverter, num) => {
+                format!("{}/read/forced_discharge/{}", inverter.datalog(), num)
+            }
             SetHold(inverter, register, _) => {
                 format!("{}/set/hold/{}", inverter.datalog(), register)
             }
+            WriteParam(inverter, register, _) => {
+                format!("{}/set/param/{}", inverter.datalog(), register)
+            }
+            SetAcChargeTime(inverter, num, _) => {
+                format!("{}/set/ac_charge/{}", inverter.datalog(), num)
+            }
+            SetChargePriorityTime(inverter, num, _) => {
+                format!("{}/set/charge_priority/{}", inverter.datalog(), num)
+            }
+            SetForcedDischargeTime(inverter, num, _) => {
+                format!("{}/set/forced_discharge/{}", inverter.datalog(), num)
+            }
             AcCharge(inverter, _) => format!("{}/set/ac_charge", inverter.datalog()),
+            ChargePriority(inverter, _) => format!("{}/set/charge_priority", inverter.datalog()),
             ForcedDischarge(inverter, _) => format!("{}/set/forced_discharge", inverter.datalog()),
             ChargeRate(inverter, _) => format!("{}/set/charge_rate_pct", inverter.datalog()),
             DischargeRate(inverter, _) => format!("{}/set/discharge_rate_pct", inverter.datalog()),