+++ src/config.rs
@@ -71,6 +71,10 @@ pub struct HomeAssistant {
     #[serde(default = "Config::default_mqtt_homeassistant_sensors")]
     #[serde_as(as = "StringWithSeparator::<CommaSeparator, String>")]
     pub sensors: Vec<String>,
+
+    #[serde(default = "Config::default_mqtt_homeassistant_switches")]
+    #[serde_as(as = "StringWithSeparator::<CommaSeparator, String>")]
+    pub switches: Vec<String>,
 }
 impl HomeAssistant {
     pub fn enabled(&self) -> bool {
@@ -83,6 +87,10 @@ impl HomeAssistant {
 
     pub fn sensors(&self) -> &Vec<String> {
         &self.sensors
+    }
+
+    pub fn switches(&self) -> &Vec<String> {
+        &self.switches
     }
 } // }}}
 
@@ -352,6 +360,7 @@ impl Config {
             enabled: Self::default_enabled(),
             prefix: Self::default_mqtt_homeassistant_prefix(),
             sensors: Self::default_mqtt_homeassistant_sensors(),
+            switches: Self::default_mqtt_homeassistant_switches(),
         }
     }
 
@@ -359,6 +368,10 @@ impl Config {
         "homeassistant".to_string()
     }
     fn default_mqtt_homeassistant_sensors() -> Vec<String> {
+        // by default, use the special-case string of "all" rather than list them all out
+        vec!["all".to_string()]
+    }
+    fn default_mqtt_homeassistant_switches() -> Vec<String> {
         // by default, use the special-case string of "all" rather than list them all out
         vec!["all".to_string()]
     }