+++ src/coordinator/mod.rs
@@ -78,6 +78,7 @@ impl Coordinator {
     }
 
     async fn process_command(&self, command: Command) -> Result<()> {
+        use commands::time_register_ops::Action;
         use lxp::packet::{Register, RegisterBit};
         use Command::*;
 
@@ -91,7 +92,34 @@ impl Coordinator {
             }
             ReadHold(inverter, register, count) => self.read_hold(inverter, register, count).await,
             ReadParam(inverter, register) => self.read_param(inverter, register).await,
+            ReadAcChargeTime(inverter, num) => {
+                self.read_time_register(inverter, Action::AcCharge(num))
+                    .await
+            }
+            ReadChargePriorityTime(inverter, num) => {
+                self.read_time_register(inverter, Action::ChargePriority(num))
+                    .await
+            }
+            ReadForcedDischargeTime(inverter, num) => {
+                self.read_time_register(inverter, Action::ForcedDischarge(num))
+                    .await
+            }
             SetHold(inverter, register, value) => self.set_hold(inverter, register, value).await,
+            WriteParam(inverter, register, value) => {
+                self.write_param(inverter, register, value).await
+            }
+            SetAcChargeTime(inverter, num, values) => {
+                self.set_time_register(inverter, Action::AcCharge(num), values)
+                    .await
+            }
+            SetChargePriorityTime(inverter, num, values) => {
+                self.set_time_register(inverter, Action::ChargePriority(num), values)
+                    .await
+            }
+            SetForcedDischargeTime(inverter, num, values) => {
+                self.set_time_register(inverter, Action::ForcedDischarge(num), values)
+                    .await
+            }
             AcCharge(inverter, enable) => {
                 self.update_hold(
                     inverter,
@@ -101,6 +129,16 @@ impl Coordinator {
                 )
                 .await
             }
+            ChargePriority(inverter, enable) => {
+                self.update_hold(
+                    inverter,
+                    Register::Register21,
+                    RegisterBit::ChargePriorityEnable,
+                    enable,
+                )
+                .await
+            }
+
             ForcedDischarge(inverter, enable) => {
                 self.update_hold(
                     inverter,
@@ -140,7 +178,7 @@ impl Coordinator {
         &self,
         inverter: config::Inverter,
         register: U,
-        count: i16,
+        count: u16,
     ) -> Result<()>
     where
         U: Into<i16>,
@@ -157,7 +195,7 @@ impl Coordinator {
         Ok(())
     }
 
-    async fn read_hold<U>(&self, inverter: config::Inverter, register: U, count: i16) -> Result<()>
+    async fn read_hold<U>(&self, inverter: config::Inverter, register: U, count: u16) -> Result<()>
     where
         U: Into<i16>,
     {
@@ -184,7 +222,58 @@ impl Coordinator {
         Ok(())
     }
 
-    async fn set_hold<U>(&self, inverter: config::Inverter, register: U, value: i16) -> Result<()>
+    async fn read_time_register(
+        &self,
+        inverter: config::Inverter,
+        action: commands::time_register_ops::Action,
+    ) -> Result<()> {
+        commands::time_register_ops::ReadTimeRegister::new(
+            self.channels.clone(),
+            inverter.clone(),
+            action,
+        )
+        .run()
+        .await
+    }
+
+    async fn write_param<U>(
+        &self,
+        inverter: config::Inverter,
+        register: U,
+        value: u16,
+    ) -> Result<()>
+    where
+        U: Into<i16>,
+    {
+        commands::write_param::WriteParam::new(
+            self.channels.clone(),
+            inverter.clone(),
+            register,
+            value,
+        )
+        .run()
+        .await?;
+
+        Ok(())
+    }
+
+    async fn set_time_register(
+        &self,
+        inverter: config::Inverter,
+        action: commands::time_register_ops::Action,
+        values: [u8; 4],
+    ) -> Result<()> {
+        commands::time_register_ops::SetTimeRegister::new(
+            self.channels.clone(),
+            inverter.clone(),
+            action,
+            values,
+        )
+        .run()
+        .await
+    }
+
+    async fn set_hold<U>(&self, inverter: config::Inverter, register: U, value: u16) -> Result<()>
     where
         U: Into<i16>,
     {