+++ tests/test_home_assistant.rs
@@ -7,8 +7,9 @@ async fn all_empty_with_no_sensors() {
 
     let mut config = Factory::example_config();
     config.mqtt.homeassistant.sensors = vec![];
+    config.mqtt.homeassistant.switches = vec![];
 
-    let r = home_assistant::Config::all(&config.inverters[0], &config.mqtt);
+    let r = home_assistant::Config::new(&config.inverters[0], &config.mqtt).all();
 
     assert!(r.is_ok());
     assert_eq!(r.unwrap(), vec![]);
@@ -20,16 +21,20 @@ async fn all_has_soc() {
 
     let mut config = Factory::example_config();
     config.mqtt.homeassistant.sensors = vec!["soc".to_owned()];
+    config.mqtt.homeassistant.switches = vec![];
 
-    let r = home_assistant::Config::all(&config.inverters[0], &config.mqtt);
+    let r = home_assistant::Config::new(&config.inverters[0], &config.mqtt).all();
 
     assert!(r.is_ok());
-    assert_eq!(r.unwrap(), vec![
-        mqtt::Message {
-            topic: "homeassistant/sensor/2222222222_soc/config".to_string(),
-            payload: "{\"device_class\":\"battery\",\"name\":\"Battery Percentage\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.soc }}\",\"unit_of_measurement\":\"%\",\"unique_id\":\"lxp_2222222222_soc\"}".to_string()
-        }
-    ]);
+    assert_eq!(
+        r.unwrap(),
+        vec![
+            mqtt::Message {
+                topic: "homeassistant/sensor/lxp_2222222222/soc/config".to_string(),
+                payload: "{\"device_class\":\"battery\",\"name\":\"Battery Percentage\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.soc }}\",\"unit_of_measurement\":\"%\",\"unique_id\":\"lxp_2222222222_soc\",\"device\":{\"manufacturer\":\"LuxPower\",\"name\":\"lxp_2222222222\",\"identifiers\":[\"lxp_2222222222\"]},\"availability\":{\"topic\":\"lxp/LWT\"}}".to_string()
+            }
+        ]
+    );
 }
 
 #[tokio::test]
@@ -38,14 +43,15 @@ async fn all_has_v_pv() {
 
     let mut config = Factory::example_config();
     config.mqtt.homeassistant.sensors = vec!["v_pv".to_owned()];
+    config.mqtt.homeassistant.switches = vec![];
 
-    let r = home_assistant::Config::all(&config.inverters[0], &config.mqtt);
+    let r = home_assistant::Config::new(&config.inverters[0], &config.mqtt).all();
 
     assert!(r.is_ok());
     assert_eq!(r.unwrap(), vec![
         mqtt::Message {
-            topic: "homeassistant/sensor/2222222222_v_pv/config".to_string(),
-            payload: "{\"device_class\":\"voltage\",\"name\":\"Voltage (PV Array)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.v_pv }}\",\"unit_of_measurement\":\"V\",\"unique_id\":\"lxp_2222222222_v_pv\"}".to_string()
+            topic: "homeassistant/sensor/lxp_2222222222/v_pv/config".to_string(),
+            payload: "{\"device_class\":\"voltage\",\"name\":\"Voltage (PV Array)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.v_pv }}\",\"unit_of_measurement\":\"V\",\"unique_id\":\"lxp_2222222222_v_pv\",\"device\":{\"manufacturer\":\"LuxPower\",\"name\":\"lxp_2222222222\",\"identifiers\":[\"lxp_2222222222\"]},\"availability\":{\"topic\":\"lxp/LWT\"}}".to_string()
         }
     ]);
 }
@@ -56,16 +62,20 @@ async fn all_has_p_pv() {
 
     let mut config = Factory::example_config();
     config.mqtt.homeassistant.sensors = vec!["p_pv".to_owned()];
+    config.mqtt.homeassistant.switches = vec![];
 
-    let r = home_assistant::Config::all(&config.inverters[0], &config.mqtt);
+    let r = home_assistant::Config::new(&config.inverters[0], &config.mqtt).all();
 
     assert!(r.is_ok());
-    assert_eq!(r.unwrap(), vec![
-        mqtt::Message {
-            topic: "homeassistant/sensor/2222222222_p_pv/config".to_string(),
-            payload: "{\"device_class\":\"power\",\"name\":\"Power (PV Array)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.p_pv }}\",\"unit_of_measurement\":\"W\",\"unique_id\":\"lxp_2222222222_p_pv\"}".to_string()
-        }
-    ]);
+    assert_eq!(
+        r.unwrap(),
+        vec![
+            mqtt::Message {
+                topic: "homeassistant/sensor/lxp_2222222222/p_pv/config".to_string(),
+                payload: "{\"device_class\":\"power\",\"name\":\"Power (PV Array)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"measurement\",\"value_template\":\"{{ value_json.p_pv }}\",\"unit_of_measurement\":\"W\",\"unique_id\":\"lxp_2222222222_p_pv\",\"device\":{\"manufacturer\":\"LuxPower\",\"name\":\"lxp_2222222222\",\"identifiers\":[\"lxp_2222222222\"]},\"availability\":{\"topic\":\"lxp/LWT\"}}".to_string()
+            }
+        ]
+    );
 }
 
 #[tokio::test]
@@ -74,14 +84,15 @@ async fn all_has_e_pv_all() {
 
     let mut config = Factory::example_config();
     config.mqtt.homeassistant.sensors = vec!["e_pv_all".to_owned()];
+    config.mqtt.homeassistant.switches = vec![];
 
-    let r = home_assistant::Config::all(&config.inverters[0], &config.mqtt);
+    let r = home_assistant::Config::new(&config.inverters[0], &config.mqtt).all();
 
     assert!(r.is_ok());
     assert_eq!(r.unwrap(), vec![
         mqtt::Message {
-            topic: "homeassistant/sensor/2222222222_e_pv_all/config".to_string(),
-            payload: "{\"device_class\":\"energy\",\"name\":\"PV Generation (All time)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"total_increasing\",\"value_template\":\"{{ value_json.e_pv_all }}\",\"unit_of_measurement\":\"kWh\",\"unique_id\":\"lxp_2222222222_e_pv_all\"}".to_string()
+            topic: "homeassistant/sensor/lxp_2222222222/e_pv_all/config".to_string(),
+            payload: "{\"device_class\":\"energy\",\"name\":\"PV Generation (All time)\",\"state_topic\":\"lxp/2222222222/inputs/all\",\"state_class\":\"total_increasing\",\"value_template\":\"{{ value_json.e_pv_all }}\",\"unit_of_measurement\":\"kWh\",\"unique_id\":\"lxp_2222222222_e_pv_all\",\"device\":{\"manufacturer\":\"LuxPower\",\"name\":\"lxp_2222222222\",\"identifiers\":[\"lxp_2222222222\"]},\"availability\":{\"topic\":\"lxp/LWT\"}}".to_string()
         }
     ]);
 }