+++ tests/test_mqtt_message.rs
@@ -46,6 +46,50 @@ async fn for_hold_single() {
 }
 
 #[tokio::test]
+async fn for_hold_21() {
+    common_setup();
+
+    let inverter = Factory::inverter();
+
+    let packet = lxp::packet::TranslatedData {
+        datalog: inverter.datalog(),
+        device_function: lxp::packet::DeviceFunction::ReadHold,
+        inverter: inverter.serial(),
+        register: 21,
+        values: vec![12, 34],
+    };
+
+    assert_eq!(
+        mqtt::Message::for_hold(packet).unwrap(),
+        vec![mqtt::Message { topic: "2222222222/hold/21".to_owned(), payload: "8716".to_owned() },
+             mqtt::Message { topic: "2222222222/hold/21/bits".to_owned(), payload: "{\"eps_en\":\"OFF\",\"ovf_load_derate_en\":\"OFF\",\"drms_en\":\"ON\",\"lvrt_en\":\"ON\",\"anti_island_en\":\"OFF\",\"neutral_detect_en\":\"OFF\",\"grid_on_power_ss_en\":\"OFF\",\"ac_charge_en\":\"OFF\",\"sw_seamless_en\":\"OFF\",\"set_to_standby\":\"ON\",\"forced_discharge_en\":\"OFF\",\"charge_priority_en\":\"OFF\",\"iso_en\":\"OFF\",\"gfci_en\":\"ON\",\"dci_en\":\"OFF\",\"feed_in_grid_en\":\"OFF\"}".to_owned() }
+        ]
+    );
+}
+
+#[tokio::test]
+async fn for_hold_110() {
+    common_setup();
+
+    let inverter = Factory::inverter();
+
+    let packet = lxp::packet::TranslatedData {
+        datalog: inverter.datalog(),
+        device_function: lxp::packet::DeviceFunction::ReadHold,
+        inverter: inverter.serial(),
+        register: 110,
+        values: vec![9, 4],
+    };
+
+    assert_eq!(
+        mqtt::Message::for_hold(packet).unwrap(),
+        vec![mqtt::Message { topic: "2222222222/hold/110".to_owned(), payload: "1033".to_owned() },
+             mqtt::Message { topic: "2222222222/hold/110/bits".to_owned(), payload: "{\"ub_pv_grid_off_en\":\"ON\",\"ub_run_without_grid\":\"OFF\",\"ub_micro_grid_en\":\"OFF\"}".to_owned() }
+        ]
+    );
+}
+
+#[tokio::test]
 async fn for_hold_multi() {
     common_setup();